(update) renamed paths to reflect new merges
