+++ client/reducers/index.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import formReducer from './formReducer';
 
 import travelReducer from './travelReducer';
-import formReducer from './formReducer';
 
 const combinedReducers = combineReducers({
   trips: travelReducer,
